pyproject.toml
[project]
name = "hoistwaywatch"
version = "0.1.0"
description = "Local-first hoistway awareness: vision -> rules -> alerts."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [{ name = "HoistwayWatch contributors" }]
dependencies = [
  "pydantic>=2.10.0",
  "pyyaml>=6.0.2",
  "nats-py>=2.9.0",
  "opencv-python-headless>=4.10.0.84",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.8.0",
]

[project.scripts]
hw-capture = "hoistwaywatch.capture.cli:main"
hw-vision = "hoistwaywatch.vision.cli:main"
hw-rules = "hoistwaywatch.rules.cli:main"
hw-alerts = "hoistwaywatch.alerts.cli:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel>=0.42.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]

